hisa-vm/pyproject.toml
[project]
name = "hisa-vm"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.6.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
